requirements.txt
streamlit==1.34.0
google-generativeai==0.3.2
httpx==0.26.0
python-dotenv==1.0.1
